+++ .vscode/settings.json
@@ -2,6 +2,7 @@
     "files.associations": {
         "cstdio": "c",
         "errno.h": "c",
-        "time.h": "c"
+        "time.h": "c",
+        "stat.h": "c"
     }
 }
